+++ Main-Saw-Fence-ClearCore/MechanismClasses.h
@@ -25,13 +25,13 @@ public:
     virtual int GetMotorProgInputRes() const = 0;
     virtual int GetMaxAccel() const = 0;
     virtual int GetMaxVel() const = 0;
-    virtual double CalculateStepsPerUnit() const = 0;
+    virtual float CalculateStepsPerUnit() const = 0;
 
     // Virtual getters for mechanism-specific parameters and their units.
     // Return 0.0 by default for types where the parameter is not applicable.
-    virtual double GetPulleyDiameter() const { return 0.0; }
-    virtual double GetLeadscrewPitch() const { return 0.0; }
-    virtual double GetPinionDiameter() const { return 0.0; }
+    virtual float GetPulleyDiameter() const { return 0.0; }
+    virtual float GetLeadscrewPitch() const { return 0.0; }
+    virtual float GetPinionDiameter() const { return 0.0; }
     virtual UnitType GetParamUnit() const = 0; // Pure virtual getter for unit type
 
     // Virtual destructor for proper polymorphic deletion
@@ -44,22 +44,22 @@ private:
     int motorProgInputRes;
     int maxAccel;
     int maxVel;
-    double gearboxReduction;
-    double pulleyDiameter; // Specific parameter for belt drive
+    float gearboxReduction;
+    float pulleyDiameter; // Specific parameter for belt drive
     UnitType pulleyDiameterUnit; // Unit for pulleyDiameter
 
 public:
     // Constructor declaration
-    BeltMechanism(int res, int accel, int vel, double diameter, double gearboxReduction, UnitType unit);
+    BeltMechanism(int res, int accel, int vel, float diameter, float gearboxReduction, UnitType unit);
 
     // Implementations of the pure virtual functions
     int GetMotorProgInputRes() const override;
     int GetMaxAccel() const override;
     int GetMaxVel() const override;
-    double CalculateStepsPerUnit() const override;
+    float CalculateStepsPerUnit() const override;
 
     // Override specific getter for pulley diameter and its unit
-    double GetPulleyDiameter() const override;
+    float GetPulleyDiameter() const override;
     UnitType GetParamUnit() const override;
 };
 
@@ -69,22 +69,22 @@ private:
     int motorProgInputRes;
     int maxAccel;
     int maxVel;
-    double gearboxReduction;
-    double leadscrewPitch; // Specific parameter for leadscrew
+    float gearboxReduction;
+    float leadscrewPitch; // Specific parameter for leadscrew
     UnitType leadscrewPitchUnit; // Unit for leadscrewPitch
 
 public:
     // Constructor declaration
-    LeadscrewMechanism(int res, int accel, int vel, double pitch, double gearboxReduction, UnitType unit);
+    LeadscrewMechanism(int res, int accel, int vel, float pitch, float gearboxReduction, UnitType unit);
 
     // Implementations of the pure virtual functions
     int GetMotorProgInputRes() const override;
     int GetMaxAccel() const override;
     int GetMaxVel() const override;
-    double CalculateStepsPerUnit() const override;
+    float CalculateStepsPerUnit() const override;
 
     // Override specific getter for leadscrew pitch and its unit
-    double GetLeadscrewPitch() const override;
+    float GetLeadscrewPitch() const override;
     UnitType GetParamUnit() const override;
 };
 
@@ -94,22 +94,22 @@ private:
     int motorProgInputRes;
     int maxAccel;
     int maxVel;
-    double gearboxReduction;
-    double pinionDiameter; // Specific parameter for rack and pinion
+    float gearboxReduction;
+    float pinionDiameter; // Specific parameter for rack and pinion
     UnitType pinionDiameterUnit; // Unit for pinionDiameter
 
 public:
     // Constructor declaration
-    RackAndPinionMechanism(int res, int accel, int vel, double diameter, double gearboxReduction, UnitType unit);
+    RackAndPinionMechanism(int res, int accel, int vel, float diameter, float gearboxReduction, UnitType unit);
 
     // Implementations of the pure virtual functions
     int GetMotorProgInputRes() const override;
     int GetMaxAccel() const override;
     int GetMaxVel() const override;
-    double CalculateStepsPerUnit() const override;
+    float CalculateStepsPerUnit() const override;
 
     // Override specific getter for pinion diameter and its unit
-    double GetPinionDiameter() const override;
+    float GetPinionDiameter() const override;
     UnitType GetParamUnit() const override;
 };
